Update run.rb output
