+++ run.rb
@@ -3,8 +3,7 @@
 
 require 'dotenv/load'
 require_relative 'config/labels'
-require_relative 'src/lib/delete'
-require_relative 'src/lib/add'
+require_relative 'src/lib/labels'
 
 def run
   # Delete existing labels
@@ -12,6 +11,7 @@ def run
 
   # Add new labels set in LABELS (config/labels.rb)
   Labels.add(LABELS)
+  puts "All labels for repo '#{ENV.fetch('GITHUB_REPO')}' were updated!"
 end
 
 run if $PROGRAM_NAME == __FILE__
